+++ content/template/detailTemplate/template_A.php
@@ -306,7 +306,7 @@ $showEventVenue = get_option('eduadmin-showEventVenueName', false);
 			<div class="eventItem">
 				<div class="eventDate<?php echo $groupByCityClass; ?>">
 					<?php echo isset($eventDates[$ev->EventID]) ? GetLogicalDateGroups($eventDates[$ev->EventID]) : GetOldStartEndDisplayDate($ev->PeriodStart, $ev->PeriodEnd); ?>
-					<?php echo (!isset($eventDates[$ev->EventID]) ? ", " . date("H:i", strtotime($ev->PeriodStart)) .' - ' . date("H:i", strtotime($ev->PeriodEnd)) : ""); ?>
+					<?php echo (!isset($eventDates[$ev->EventID]) ? "<span class=\"eventTime\">, " . date("H:i", strtotime($ev->PeriodStart)) .' - ' . date("H:i", strtotime($ev->PeriodEnd)) . "</span>" : ""); ?>
 				</div>
 				<?php if(!$groupByCity) { ?>
 				<div class="eventCity">